requirements.txt
# SmartLife Bot - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Bot Framework
python-telegram-bot==20.7
httpx<0.27
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23

# Google AI Services (Free APIs)
google-generativeai==0.3.1        # Gemini API (FREE 1M tokens/day)
google-cloud-speech==2.21.0       # Speech-to-Text (FREE 60min/month)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0 # Calendar API (FREE)

# Task Scheduling (No Redis needed!)
apscheduler==3.10.4

# Data Visualization
matplotlib==3.8.2
pandas==2.1.4

# Utilities
pytz==2023.3
requests==2.32.4


# Google Calendar API (NEW for Phase 3)
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
# Google Gemini AI
google-generativeai==0.3.1

# Date/Time handling
python-dateutil==2.8.2